backend/requirements.txt
# 留学双边信息平台 - 后端依赖包
# 更新日期: 2024-01-01
# Python 版本要求: >=3.11

# 核心Web框架
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0

# 数据验证与设置
pydantic[email]==2.11.7
pydantic-settings==2.10.1

# 数据库连接
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.36
supabase==2.17.0

# 认证与安全
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# 环境配置
python-dotenv==1.0.1

# HTTP客户端
requests==2.32.4

# 文件上传支持
python-multipart==0.0.19

# AI Agent相关
langchain==0.3.8
langchain-openai==0.3.1
langchain-community==0.3.8
langchain-core==0.3.19
langgraph==0.2.51
tavily-python==0.7.10
duckduckgo-search==7.0.3

# 知识库与向量存储
chromadb==0.6.2
unstructured[docx,pdf]==0.17.5
pypdf==5.2.0

# Web界面
streamlit==1.41.1

# 开发与测试
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1